fix: address cell-grid review findings (slot width, atlas geometry source)
From PR #956 review:
- GlyphAtlas slot indices stay usize end to end -- the u16 in the char
  map would wrap past 65,535 baked glyphs and silently serve unrelated
  tiles on later lookups.
- The atlas follows the SNAPSHOT's cell geometry, not the startup
  config: build_scene now rebuilds the atlas (ensure_atlas) whenever the
  emulator's cell height or the display density diverges from what the
  tiles were baked at, so a config/emulator mismatch can no longer scale
  or clip every glyph. set_density becomes a plain field write; the
  rebuild and the program recompile both key off the snapshot.
- Documented the one-font contract on the atlas slot map: keyed by char
  alone by design; font switches build a new atlas through the same path
  that rebuilds on size/density changes.

+++ pixelflow-graphics/src/fonts/atlas.rs
@@ -52,7 +52,12 @@ pub struct GlyphAtlas {
     capacity: usize,
     /// `char` → slot. `None` records "font has no glyph" so a missing char
     /// is looked up (and logged by the caller) once, not per frame.
-    slots: HashMap<char, Option<u16>>,
+    ///
+    /// An atlas is bound to ONE font: the map is keyed by `char` alone, so
+    /// feeding `uv` different fonts would serve stale tiles. Callers that
+    /// switch fonts build a new atlas (core-term does, via the same path
+    /// that rebuilds on cell-size and density changes).
+    slots: HashMap<char, Option<usize>>,
     /// Next unassigned slot. Slot 0 is the reserved blank.
     next_slot: usize,
     /// Point-space glyph square (the cell height, by convention).
@@ -156,7 +161,7 @@ impl GlyphAtlas {
     pub fn uv(&mut self, font: &Font, ch: char) -> (f32, f32) {
         if let Some(&slot) = self.slots.get(&ch) {
             return match slot {
-                Some(s) => self.slot_origin(s as usize),
+                Some(s) => self.slot_origin(s),
                 None => self.blank_uv(),
             };
         }
@@ -180,7 +185,7 @@ impl GlyphAtlas {
         }
         let slot = self.next_slot;
         self.next_slot += 1;
-        self.slots.insert(ch, Some(slot as u16));
+        self.slots.insert(ch, Some(slot));
 
         let (u, v) = self.slot_origin(slot);
         let (u, v) = (u as usize, v as usize);
